+++ src/configToSchemas/index.js
@@ -1,14 +1,19 @@
-import * as builtInTypes from './types';
-import typePlugins from './typePlugins';
-import commonPlugins from './commonPlugins';
 import { mapValues } from 'lodash';
+import commonPlugins from './commonPlugins';
+import typePlugins from './typePlugins';
+import * as builtInTypes from './types';
 
 export default (props) => {
   const types = {
     ...mapValues(builtInTypes, (type) => type.default),
     ...props.extraTypes,
   };
-  const typeConverter = types[props.config.type];
+
+  const typeConverter =
+    typeof props.config.type === 'function'
+      ? props.config.type
+      : types[props.config.type];
+
   if (!typeConverter) {
     // eslint-disable-next-line no-console
     console.error(`Config type "${props.config.type}" is not implemented.`);